+++ src/pages/StationInfo.jsx
@@ -1,5 +1,12 @@
+import Header from "@/components/Header";
+
 function StationInfo() {
-  return <div>2-2 回收站點詳細資訊頁</div>;
+  return (
+    <>
+      <Header />
+      <main>2-2 回收站點詳細資訊頁</main>
+    </>
+  );
 }
 
 export default StationInfo;
